M8-Full_stack/flask-blog-master-day4/requirements.txt
Authlib==0.15.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
Flask-CKEditor==0.4.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pycparser==2.21
python-dotenv==0.19.1
python-slugify==5.0.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
text-unidecode==1.3
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
